requirements.txt
tweepy>=4.14.0
google-genai>=0.3.0
